Move visitors to a shared module
